+++ src/main/java/oscar/oscarPrevention/reports/ChildImmunizationReport.java
@@ -252,7 +252,7 @@ public class ChildImmunizationReport implements PreventionReport{
           String percentStr = "0";
           double eligible = list.size() - inList - dontInclude;
           log.debug("eligible "+eligible+" done "+done);
-          if (eligible != 0){
+          if ((int)eligible != 0){
              double percentage = ( done / eligible ) * 100;
              log.debug("in percentage  "+percentage   +" "+( done / eligible));
              percentStr = ""+Math.round(percentage);
